Bio2/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
ultralytics==8.2.0
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
loguru==0.7.2
sqlalchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pillow==10.4.0
requests==2.32.3